+++ compiler/src/it/producers-functional-tests/src/main/java/producerstest/DependedProductionComponent.java
@@ -13,22 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package test;
+package producerstest;
 
-import com.google.common.base.Ascii;
-import com.google.common.collect.ImmutableList;
-import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
-import dagger.producers.ProducerModule;
-import dagger.producers.Produces;
+import dagger.producers.ProductionComponent;
 
-import java.util.List;
-
-@ProducerModule
-final class DependedProducerModule {
-
-  @Produces
-  int produceNumberOfGreetings() {
-    return 2;
-  }
+@ProductionComponent(modules = DependedProducerModule.class)
+interface DependedProductionComponent {
+  ListenableFuture<Integer> numGreetings();
 }
+
