+++ compiler/src/it/producers-functional-tests/src/main/java/producerstest/DependentComponent.java
@@ -13,12 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package test;
+package producerstest;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import dagger.producers.ProductionComponent;
 
-@ProductionComponent(modules = ResponseProducerModule.class)
-interface SimpleComponent {
-  ListenableFuture<Response> response();
+import java.util.List;
+
+@ProductionComponent(
+    modules = DependentProducerModule.class,
+    dependencies = {DependedComponent.class, DependedProductionComponent.class})
+interface DependentComponent {
+  ListenableFuture<List<String>> greetings();
 }
