+++ compiler/src/it/producers-functional-tests/src/main/java/producerstest/ResponseModule.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package test;
+package producerstest;
 
 import dagger.Module;
 import dagger.Provides;
