+++ compiler/src/it/producers-functional-tests/src/main/java/producerstest/SimpleComponent.java
@@ -13,23 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package test;
+package producerstest;
 
-import com.google.common.base.Ascii;
-import com.google.common.collect.ImmutableList;
-import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
-import dagger.producers.ProducerModule;
-import dagger.producers.Produces;
+import dagger.producers.ProductionComponent;
 
-import java.util.List;
-
-@ProducerModule
-final class DependentProducerModule {
-  @Produces
-  ListenableFuture<List<String>> greetings(Integer numGreetings, String greeting) {
-    List<String> greetings = ImmutableList.of(
-        String.valueOf(numGreetings), greeting, Ascii.toUpperCase(greeting));
-    return Futures.immediateFuture(greetings);
-  }
+@ProductionComponent(modules = ResponseProducerModule.class)
+interface SimpleComponent {
+  ListenableFuture<Response> response();
 }
