+++ compiler/src/it/producers-functional-tests/src/test/java/producerstest/MultibindingTest.java
@@ -13,7 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package test;
+package producerstest;
 
 import com.google.common.collect.Iterables;
 import com.google.common.util.concurrent.MoreExecutors;
