+++ compiler/src/it/producers-functional-tests/src/test/java/producerstest/ProducerFactoryTest.java
@@ -13,7 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package test;
+package producerstest;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.fail;
